refactor: Change js type module in signin
