+++ js/test.js
@@ -1,0 +1,6 @@
+import { INPUT_TYPE } from './input.js';
+
+export const USER_TEST = {
+  [INPUT_TYPE.EMAIL]: 'test@codeit.com',
+  [INPUT_TYPE.PASSWORD]: 'codeit101',
+};
