+++ js/validation.js
@@ -1,0 +1,4 @@
+export function validateEmail(email) {
+  const emailRegex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/i;
+  return emailRegex.test(email);
+}
