+++ signup/index.html
@@ -28,7 +28,7 @@
             </div>
           </div>
         </div>
-        <form id="sign-form" action="/folder" method="post">
+        <form id="signup-form" action="/folder" method="post">
           <div class="inputs">
             <div class="input-container">
               <label for="email">이메일</label>
@@ -103,6 +103,6 @@
         </div>
       </div>
     </div>
-    <script src="/js/sign.js"></script>
+    <script type="module" src="/js/signup.js"></script>
   </body>
 </html>
